requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
altgraph==0.17.4
    # via
    #   -r requirements.in
    #   macholib
    #   pyinstaller
beautifulsoup4==4.12.3
    # via -r requirements.in
build==1.2.2.post1
    # via
    #   -r requirements.in
    #   pip-tools
cachetools==5.5.0
    # via
    #   -r requirements.in
    #   google-auth
certifi==2024.12.14
    # via
    #   -r requirements.in
    #   requests
charset-normalizer==3.4.1
    # via
    #   -r requirements.in
    #   requests
click==8.1.8
    # via
    #   -r requirements.in
    #   pip-tools
coverage==7.6.10
    # via -r requirements.in
google-api-core==2.24.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-api-python-client==2.156.0
    # via -r requirements.in
google-auth==2.37.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.2.1
    # via -r requirements.in
googleapis-common-protos==1.66.0
    # via
    #   -r requirements.in
    #   google-api-core
httplib2==0.22.0
    # via
    #   -r requirements.in
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.10
    # via
    #   -r requirements.in
    #   requests
macholib==1.16.3
    # via
    #   -r requirements.in
    #   pyinstaller
oauthlib==3.2.2
    # via
    #   -r requirements.in
    #   requests-oauthlib
packaging==24.2
    # via
    #   -r requirements.in
    #   build
    #   pyinstaller
    #   pyinstaller-hooks-contrib
pip-tools==7.4.1
    # via -r requirements.in
proto-plus==1.25.0
    # via
    #   -r requirements.in
    #   google-api-core
protobuf==5.29.2
    # via
    #   -r requirements.in
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.1
    # via
    #   -r requirements.in
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via
    #   -r requirements.in
    #   google-auth
pyinstaller==6.11.1
    # via -r requirements.in
pyinstaller-hooks-contrib==2024.11
    # via
    #   -r requirements.in
    #   pyinstaller
pyparsing==3.2.1
    # via
    #   -r requirements.in
    #   httplib2
pyproject-hooks==1.2.0
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
python-dateutil # Added
requests==2.32.3
    # via
    #   -r requirements.in
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via
    #   -r requirements.in
    #   google-auth-oauthlib
rsa==4.9
    # via
    #   -r requirements.in
    #   google-auth
soupsieve==2.6
    # via
    #   -r requirements.in
    #   beautifulsoup4
tabulate==0.9.0
    # via -r requirements.in
uritemplate==4.1.1
    # via
    #   -r requirements.in
    #   google-api-python-client
urllib3==2.3.0
    # via
    #   -r requirements.in
    #   requests
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
